cleanup: remove ReefGuideApiService, CriteriaRequest
