--- packages/app/src/app/location-selection/reef-guide-api.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ReefGuideApiService } from './reef-guide-api.service';
-
-describe('SelectionApiService', () => {
-  let service: ReefGuideApiService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ReefGuideApiService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
